Clean up mocks, add basic model query test
