+++ Tests/AppTests/Mocks/TestUrls+mock.swift
@@ -1,74 +1,3 @@
-import Vapor
-
-
-#if DEBUG
-
-extension AppEnvironment {
-    static let mock: Self = .init(
-        date: Date.init,
-        fetchMasterPackageList: { _ in
-            .just(value: ["https://github.com/finestructure/Gala",
-                          "https://github.com/finestructure/SwiftPMLibrary-Server"].urls)
-        },
-        fetchMetadata: { _, _ in .just(value: .mock) },
-        fileManager: .mock,
-        githubToken: { nil },
-        reportError: { _, _, _ in .just(value: ()) },
-        rollbarToken: { nil },
-        shell: .mock
-    )
-
-    static let e2eTestingShort: Self = .init(
-        date: Date.init,
-        fetchMasterPackageList: { _ in .just(value: testUrls) },
-        fetchMetadata: { _, pkg in .just(value: .mock(for: pkg)) },
-        fileManager: .mock,
-        githubToken: { nil },
-        reportError: { _, _, _ in .just(value: ()) },
-        rollbarToken: { nil },
-        shell: .mock
-    )
-
-    static let e2eTestingFull: Self = .init(
-        date: Date.init,
-        fetchMasterPackageList: liveFetchMasterPackageList,
-        fetchMetadata: { _, pkg in .just(value: .mock(for: pkg)) },
-        fileManager: .mock,
-        githubToken: { nil },
-        reportError: { _, _, _ in .just(value: ()) },
-        rollbarToken: { nil },
-        shell: .mock
-    )
-}
-
-
-extension FileManager {
-    static let mock = Self.mock(fileExists: true)
-    static func mock(fileExists: Bool) -> Self {
-        .init(
-            checkoutsDirectory: { DirectoryConfiguration.detect().workingDirectory + "SPI-checkouts" },
-            createDirectory: { path, _, _ in
-                print("ℹ️ MOCK: imagine we're creating a directory at path: \(path)")
-            },
-            fileExists: { path in
-                print("ℹ️ MOCK: file at \(path) exists")
-              
-                return fileExists
-            },
-            workingDirectory: { DirectoryConfiguration.detect().workingDirectory }
-        )
-    }
-}
-
-
-extension Shell {
-    static let mock: Self = .init(run: { cmd, path in
-        print("ℹ️ MOCK: imagine we're running \(cmd) at path: \(path)")
-        return ""
-    })
-}
-
-
 let testUrls100 = [
     "https://github.com/mw99/OhhAuth.git",
     "https://github.com/quanvo87/GroupWork.git",
@@ -174,32 +103,3 @@ let testUrls100 = [
 
 
 let testUrls = Array(testUrls100[..<20])
-
-
-extension Array where Element == String {
-    var urls: [URL] { compactMap(URL.init(string:)) }
-}
-
-
-extension Github.Metadata {
-    static let mock: Self = .init(
-        defaultBranch: "master",
-        description: "desc",
-        forksCount: 1,
-        license: .init(key: "mit"),
-        stargazersCount: 2,
-        parent: nil
-    )
-
-    static func mock(for package: Package) -> Self {
-        // populate with some mock data derived from the package
-        .init(defaultBranch: "master",
-              description: "This is package " + package.url,
-              forksCount: package.url.count,
-              license: .init(key: "mit"),
-              stargazersCount: package.url.count + 1,
-              parent: nil)
-    }
-}
-
-#endif
